Sum of Every Fourth Number
In this program, a user is prompted to input two positive numbers, a start number and an end number. The numbers must be below 34, but the user can choose which one is bigger. Inputting a non-numeric character other than "q" will not produce anything and the program will wait for a number. The program will find the sum of every fourth number between the two inputs plus the starting number and display the result right after the user is done. If the user wishes to quit the program, they can simply tap the "q" button and will be met with a goodbye message.