Illustrate NCSU Logo Using Assembly
This program will illustrate the NCSU letters in red onto the display. The user then has the option of changing the colors (red, green, blue, yellow, white) by using the keyboard. If the user chooses to quit, they can simply press "q" and this will display a goodbye message and halt the program. Enjoy!